Stop providing defaults in NewCommand
